requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alembic==1.7.1
    # via -r requirements.in
anyio==3.6.1
    # via httpcore
asgiref==3.5.2
    # via uvicorn
asyncpg==0.24.0
    # via -r requirements.in
attrs==22.1.0
    # via pytest
bcrypt==3.2.2
    # via -r requirements.in
black==22.6.0
    # via -r requirements.in
build==0.8.0
    # via pip-tools
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
cffi==1.15.1
    # via bcrypt
click==8.1.3
    # via
    #   black
    #   pip-tools
    #   uvicorn
dnspython==2.2.1
    # via email-validator
email-validator==1.0.3
    # via -r requirements.in
fastapi==0.68.1
    # via -r requirements.in
flake8==4.0.1
    # via -r requirements.in
greenlet==1.1.3
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   rfc3986
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements.in
mako==1.2.2
    # via alembic
markupsafe==2.1.1
    # via mako
mccabe==0.6.1
    # via flake8
mypy==0.971
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via
    #   build
    #   pytest
pathspec==0.9.0
    # via black
pep517==0.13.0
    # via build
pip-tools==6.8.0
    # via -r requirements.in
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via cffi
pydantic==1.9.2
    # via
    #   -r requirements.in
    #   fastapi
    #   sqlmodel
pyflakes==2.4.0
    # via flake8
pyjwt==1.7.1
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.19.0
    # via -r requirements.in
python-dotenv==0.10.4
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via python-multipart
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.40
    # via
    #   -r requirements.in
    #   alembic
    #   sqlmodel
sqlalchemy2-stubs==0.0.2a26
    # via sqlmodel
sqlmodel==0.0.4
    # via -r requirements.in
starlette==0.14.2
    # via fastapi
tomli==2.0.1
    # via
    #   black
    #   build
    #   mypy
    #   pep517
    #   pytest
typing-extensions==4.3.0
    # via
    #   mypy
    #   pydantic
    #   sqlalchemy2-stubs
uvicorn==0.15.0
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
